requirements.txt
astroid==1.6.1
autopep8==1.3.4
certifi==2018.1.18
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
Django==1.11
django-filter==1.1.0
django-rest-swagger==2.1.2
djangorestframework==3.7.7
flake8==3.5.0
idna==2.6
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
Markdown==2.6.11
MarkupSafe==1.0
mccabe==0.6.1
openapi-codec==1.3.2
pycodestyle==2.3.1
pyflakes==1.6.0
pylint==1.8.2
pytz==2018.3
requests==2.18.4
simplejson==3.13.2
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
wrapt==1.10.11
